backend/requirements.txt
asgiref
channels
daphne
Django
django-channels
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv